Created some great finctionality
Provided a method to generate terrain not only around a player when they move, but around enemies as well! This allows them to perform checks against terrain that exists only around them in case the player hasn't generated the terrain just yet.

Added custom data to the tile map as names of tiles...I guess I can make a bool instead of a string and just make it 1 for walkable or 0 for not walkable.